requirements.txt
requests>=2.25.0
python-logging-loki>=0.3.1
